ENG-10259: rename files and functions to be more consistent with java client naming
